internal: Move all action creators to common directory
