+++ packages/rest-hooks/src/state/actions/createReceive.ts
@@ -3,13 +3,13 @@ import {
   FetchAction,
   ReceiveAction,
   PurgeAction,
-  ResponseActions,
+  ReceiveTypes,
 } from 'rest-hooks/types';
 import { RECEIVE_TYPE, RECEIVE_MUTATE_TYPE } from 'rest-hooks/actionTypes';
 
-import { RPCAction } from '..';
+import { RPCAction } from '../..';
 
-export function createReceive<
+export default function createReceive<
   Payload extends object | string | number = object | string | number,
   S extends Schema = any
 >(
@@ -39,7 +39,9 @@ export function createReceive<
     expiresAt: now + expiryLength,
   };
   if (
-    ([RECEIVE_TYPE, RECEIVE_MUTATE_TYPE] as string[]).includes(responseType)
+    ([RECEIVE_TYPE, RECEIVE_MUTATE_TYPE] as ReceiveTypes[]).includes(
+      responseType,
+    )
   ) {
     meta.updaters = updaters;
   }
@@ -47,29 +49,5 @@ export function createReceive<
     type: responseType as any,
     payload: data,
     meta,
-  };
-}
-
-export function createReceiveError<S extends Schema = any>(
-  error: any,
-  { schema, url, responseType, options = {} }: FetchAction<any, S>['meta'],
-  { errorExpiryLength }: { errorExpiryLength: number },
-): ResponseActions {
-  const expiryLength = options.errorExpiryLength ?? errorExpiryLength;
-  /* istanbul ignore next */
-  if (process.env.NODE_ENV === 'development' && expiryLength < 0) {
-    throw new Error('Negative errorExpiryLength are not allowed.');
-  }
-  const now = Date.now();
-  return {
-    type: responseType as any,
-    payload: error,
-    meta: {
-      schema,
-      url,
-      date: now,
-      expiresAt: now + expiryLength,
-    },
-    error: true,
   };
 }
